Remove unnecessary db connection object attached to every model instance
Every instance had to have a dedicated db connection object instance only used by the save method in basemodel.

The main motive for this was so each instance would only use one connection and to occupy the db too much.

But this is would clog up the db more, as the connection would not be closed until model instance was deleted.

For a comment system, 2000 comments would mean 2000 connections just to receive info which would have a huge performance hit.